Extracting common interface, stub out OSV provider
