+++ models/interfaces.go
@@ -1,0 +1,6 @@
+package models
+
+type Provider interface {
+	Info() string
+	Scan(purls []string, username, token string) (coordinateResponses []Package, err error)
+}
