+++ providers/ossindex/OSSIndex_test.go
@@ -1,0 +1,12 @@
+package ossindex
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+)
+
+func TestInfo(t *testing.T) {
+	info := Info()
+	assert.Equal(t, "Sonatype OSS Index (https://ossindex.sonatype.org)", info)
+}
