+++ providers/osv/osv.go
@@ -1,0 +1,11 @@
+package osv
+
+// const OSSINDEX_URL = "https://ossindex.sonatype.org/api/v3/authorized/component-report"
+
+// type CoordinateRequest struct {
+// 	Coordinates []string `json:"coordinates"`
+// }
+
+func Info() string {
+	return "OSV Vulnerability Database (https://osv.dev)"
+}
