+++ providers/osv/osv_test.go
@@ -1,0 +1,12 @@
+package osv
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+)
+
+func TestInfo(t *testing.T) {
+	info := Info()
+	assert.Equal(t, "OSV Vulnerability Database (https://osv.dev)", info)
+}
